requirements.txt
hydra-core==1.3.2
matplotlib==3.5.2
numpy==1.21.5
scikit_learn==1.2.2
seaborn==0.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.10
